+++ crates/compiler/src/selector/mod.rs
@@ -40,6 +40,7 @@ impl Selector {
                 Some(parent.0.clone())
             },
             implicit_parent,
+            false,
         )?))
     }